[#4] Fix : Response, Exception 처리 코드리뷰 반영
