+++ src/main/java/com/pcb/audy/global/exception/GlobalException.java
@@ -1,0 +1,14 @@
+package com.pcb.audy.global.exception;
+
+import com.pcb.audy.global.response.ResultCode;
+import lombok.Getter;
+
+@Getter
+public class GlobalException extends RuntimeException {
+
+    private final ResultCode resultCode;
+
+    public GlobalException(ResultCode resultCode) {
+        this.resultCode = resultCode;
+    }
+}
