+++ src/main/java/com/pcb/audy/global/response/BasicResponse.java
@@ -1,11 +1,10 @@
-package com.pcb.audy.global.response.result;
+package com.pcb.audy.global.response;
 
 import java.io.Serializable;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.springframework.http.HttpStatus;
 
 @Getter
 @Builder
@@ -13,14 +12,12 @@ import org.springframework.http.HttpStatus;
 @AllArgsConstructor
 public class BasicResponse<T> implements Serializable {
 
-    private final HttpStatus status;
-    private final String code;
+    private final Integer code;
     private final String message;
     private final Object data;
 
     public static <T> BasicResponse<T> success() {
         return BasicResponse.<T>builder()
-                .status(ResultCode.SUCCESS.getStatus())
                 .code(ResultCode.SUCCESS.getCode())
                 .message(ResultCode.SUCCESS.getMessage())
                 .build();
@@ -28,10 +25,16 @@ public class BasicResponse<T> implements Serializable {
 
     public static <T> BasicResponse<T> success(T data) {
         return BasicResponse.<T>builder()
-                .status(ResultCode.SUCCESS.getStatus())
                 .code(ResultCode.SUCCESS.getCode())
                 .message(ResultCode.SUCCESS.getMessage())
                 .data(data)
+                .build();
+    }
+
+    public static <T> BasicResponse<T> error(ResultCode resultCode) {
+        return BasicResponse.<T>builder()
+                .code(resultCode.getCode())
+                .message(resultCode.getMessage())
                 .build();
     }
 }
